==> stock_market_analysis/Random Forest/3_FeatureTransformation.ipynb <==
import os
import pandas as pd
import numpy as np
import yfinance as yf
import talib as ta
import matplotlib.pyplot as plt

pd.set_option("display.max_columns", None)

final_df = pd.read_csv(os.path.join('Data', '2_NIFTY500_5y_WithIndicator.csv'), parse_dates=True, index_col=[0])

final_df.shape

train_df = final_df.loc[final_df.index < "2023-04-01"]

# does not include past 3 months data
train_df.to_csv(os.path.join("Data", "3_Train_5y.csv"))

hold_out_df = final_df.loc[(final_df.index <= "2023-06-27") & (final_df.index >= "2023-04-01")]

hold_out_df.to_csv(os.path.join("Data", "3_HoldOut_5y.csv"))

test_df = final_df.loc[final_df.index > "2023-06-27"]

test_df.to_csv(os.path.join("Data", "3_Test_5y.csv"))
