==> stock_market_analysis/Random Forest/4_ModelBuilding.ipynb <==
import os
import pandas as pd
import numpy as np
import yfinance as yf
import talib as ta
import matplotlib.pyplot as plt

pd.set_option("display.max_columns", None)

features = [
    "div_1_price_diff_values_SMA",
    # "std_21_SMA",
    # "std_7_SMA",
    # "21_SMA",
    # "7_SMA",
    # "ratio_21_SMA",
    # "ratio_7_SMA",

    "div_1_price_diff_values_EMA",
    "div_5_price_diff_values_EMA",
    "conv_5_price_diff_values_EMA",
    # "std_21_EMA",
    # "std_7_EMA",
    # "21_EMA",
    # "7_EMA",
    # "ratio_21_EMA",
    # "ratio_7_EMA",

    "div_1_price_diff_values_DEMA",
    "conv_5_price_diff_values_DEMA",
    # "std_21_DEMA",
    # "std_7_DEMA",
    # "21_DEMA",
    # "7_DEMA",
    # "ratio_21_DEMA",
    # "ratio_7_DEMA",

    "conv_1_price_diff_values_RSI",
    "div_1_price_diff_values_RSI",
    "div_5_price_diff_values_RSI",
    "conv_5_price_diff_values_RSI",
    # "rsi14_RSI",

    "moneyFlowRatio_MFI",
    "change_MFI",
    # "div_MFI",
    "momentum_MFI",
    "negMoneyFlow_MFI",
    "posMoneyFlow_MFI",
    # "typicalPrc_MFI",
    # "moneyFlow_MFI",
    "mfi14_MFI",
    # "movingAverage_MFI",

    "div_1_price_diff_values_MACD",
    "hist_MACD",
    "conv_5_price_diff_values_MACD",
    # "macd_MACD",
    "signal_MACD",
    "diff_MACD",

    "conv_1_price_diff_values_CCI",
    "trend_CCI",
    "conv_5_price_diff_values_CCI",
    # "14_CCI",
    
    "positive_momentum_1_price_diff_BBANDS",
    "positive_momentum_5_price_diff_BBANDS",
    # "lower_BBANDS",
    # "upper_BBANDS",
    
    # "10_ROC",
    "div_ROC",
    
    "%K_STOCH",
    "%D_STOCH",
    
    # "Open",
    # "High",
    # "Low",
    "Close",
    "Volume",
    "Symbol",
    "rolling_return",
    # "rolling_ret%"
    
    "trend_7_21_SMA",
    "div_SMA",
    "conv_SMA",

    "conv_EMA",
    "trend_721_EMA",
    "div_EMA",

    "conv_DEMA",
    "trend721_DEMA",
    "div_DEMA",

    "div_RSI",
    "overbought_RSI",
    "conv_RSI",
    "oversold_RSI",

    "crossOverAbove_MFI",
    "overbought_MFI",
    "oversold_MFI",

    "conv_MACD",
    "div_MACD",

    "div_CCI",
    "conv_CCI",

    "neg_momentum_BBANDS",
    "pos_momentum_BBANDS",

    "%KCross_STOCH",
    "%DOversold_STOCH",
    "%KOversold_STOCH",
    "%DOverbought_STOCH",
    "%KOverbought_STOCH",
]

df_train = pd.read_csv(os.path.join('Data', '3_Train_5y.csv'), parse_dates = ["Date"], index_col = "Date")
df_hold = pd.read_csv(os.path.join('Data', '3_HoldOut_5y.csv'), parse_dates = ["Date"], index_col = "Date")
df_test = pd.read_csv(os.path.join('Data', '3_Test_5y.csv'), parse_dates = ["Date"], index_col = "Date")

# df_train.rename({'rolling_return': 'rolling_ret'}, axis=1)
# df_hold.rename({'rolling_return': 'rolling_ret'}, axis=1)
# df_test.rename({'rolling_return': 'rolling_ret'}, axis=1)

df = df_train[features]
df.dropna(inplace = True)

from sklearn.preprocessing import LabelEncoder

le = LabelEncoder()
le.fit(df["Symbol"])

df["Symbol"] = le.transform(df["Symbol"])

# X_train, X_test, y_train, y_test = df_train.drop(["rolling_ret"], axis = 1), df_test.drop(["rolling_ret"], axis = 1), df_train["rolling_ret"].values, df_test["rolling_ret"].values
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(df.drop(["rolling_return"], axis = 1), df["rolling_return"], stratify = df["rolling_return"])

X_train.shape, X_test.shape, y_train.shape, y_test.shape

X_train.reset_index().drop("Date", axis = 1)

import numpy as np
X_train.replace({-np.inf: -1_000_000, np.inf: 1_000_000}, inplace = True)
X_test.replace({-np.inf: -1_000_000, np.inf: 1_000_000}, inplace = True)

from sklearn.ensemble import RandomForestClassifier

rfc = RandomForestClassifier(random_state = 8)

rfc.fit(X_train, y_train)

y_pred_train = rfc.predict(X_train)

y_pred_test = rfc.predict(X_test)

rfc.score(X_test, y_test)

### HoldOut

df_hold = df_hold[features]
df_hold.shape

df_hold.dropna(inplace = True)

df_hold["Symbol"] = le.transform(df_hold["Symbol"])

y_true = df_hold["rolling_return"]
X = df_hold.drop(["rolling_return"], axis = 1)

X.replace({-np.inf: -1_000_000, np.inf: 1_000_000}, inplace = True)

y_pred = rfc.predict(X)

from sklearn.metrics import accuracy_score
accuracy_score(y_true, y_pred)

from sklearn.metrics import classification_report
# holdout set
print(classification_report(y_true, y_pred))

print(classification_report(y_train, y_pred_train))

print(classification_report(y_test, y_pred_test))

### Test

df_test = df_test[features]
df_test.shape

df_test.dropna(inplace = True)

df_test["Symbol"] = le.transform(df_test["Symbol"])

y_true_test = df_test["rolling_return"]
X_test = df_test.drop(["rolling_return"], axis = 1)

X_test.replace({-np.inf: -1_000_000, np.inf: 1_000_000}, inplace = True)

y_pred_test = rfc.predict(X_test)

from sklearn.metrics import accuracy_score
accuracy_score(y_true_test, y_pred_test)
