==> stock_market_analysis/Random Forest/6_BackTesting.ipynb <==
import os
import numpy as np
import pandas as pd

stocks = pd.read_csv(os.path.join('test.csv'))
stocks = stocks.loc[stocks['1'] >= 0.98]
stocks_list = stocks.iloc[:, 0]

period_to_check = list(pd.date_range(start='2023-08-07', periods=5, freq='B').strftime("%Y-%m-%d"))
period_to_check

df_whole = pd.read_csv(os.path.join('Data', '2_NIFTY500_5y_WithIndicator.csv'), parse_dates = ["Date"], index_col = "Date")
y_buy = df_whole.loc[(df_whole.index > min(period_to_check)) & (df_whole.index <= max(period_to_check))]
y_buy

stocks_to_buy = y_buy.loc[y_buy['Symbol'].isin(stocks_list), ['Open', 'High', 'Low', 'Close', 'Symbol']]
stocks_to_buy = stocks_to_buy.reset_index()

# https://stackoverflow.com/questions/40256338/calculating-average-true-range-atr-on-ohlc-data-with-python
def wwma(values, n):
    """
     J. Welles Wilder's EMA 
    """
    return values.ewm(alpha=1/n, adjust=False).mean()

def atr(df, n=14):
    data = df.copy()
    high = data['High']
    low = data['Low']
    close = data['Close']
    data['tr0'] = abs(high - low)
    data['tr1'] = abs(high - close.shift())
    data['tr2'] = abs(low - close.shift())
    tr = data[['tr0', 'tr1', 'tr2']].max(axis=1)
    atr = wwma(tr, n)
    return atr

STYLE = 'TARGET'
QTY_TO_BUY = 1
# SL
TARGET = 3

stocks_to_buy['day_pt_move'] = (stocks_to_buy['High'] - stocks_to_buy['Low'])
stocks_to_buy['day_pct_move'] = (((stocks_to_buy['High'] - stocks_to_buy['Low']) / stocks_to_buy['Low']) * 100)

stocks_to_buy

grp_to_buy = stocks_to_buy.groupby('Date')
day_value_list = []
for grp in grp_to_buy:


if STYLE == 'TARGET':
    stocks_to_buy.groupby()
elif STYLE == 'MAX':
    pass
elif STYLE == 'STRICT':
    pass
